WIP finishing up cleanup and getting the basic functionality
Near done with basic math, broke it again near the end though.
Added -v option to show the engine's 'work' when solving
Fixed variable read, but it will need to be changed to be read on the fly later
on in the process, perhaps even at the calculate step, and change the dLeft and
dRight vars to strings.